refactor(frontend): split API client by feature
