+++ web/src/lib/api-client/index.ts
@@ -1,0 +1,8 @@
+export * from "./types";
+export * from "./core";
+export * from "./auth";
+export * from "./channels";
+export * from "./twitch";
+export * from "./recordings";
+export * from "./youtube";
+export * from "./version";
