+++ web/src/lib/api-client/version.ts
@@ -1,0 +1,17 @@
+import { isObject, safeJson, readError, request } from "./core";
+import type { VersionResponse } from "./types";
+
+export async function getVersion(): Promise<VersionResponse> {
+  const response = await request("/api/version");
+  if (!response.ok) {
+    const payload = await safeJson(response);
+    throw new Error(readError(payload));
+  }
+
+  const payload = await safeJson(response);
+  if (!isObject(payload) || typeof payload.version !== "string") {
+    throw new Error("version payload is invalid");
+  }
+
+  return { version: payload.version };
+}
